@tolitius changes requested + handle `singleton` scheduler
 ## preface
@tolitius added some comments, which are fixed and alos taking a point

> _`funcade` has to handle the scheduling within library_

For the very same imported `tolitius/yang` library which helps with
`yang.scheduler` to run a scheduled job. Also since `funcade` handles
`reitit` / `buddy` middleware it might be attached with a
`ring-handler`. The **problem** we might have is since its `handler-fn`
gets associated with `middleware`, when it gets `attached` to
`http-server` the `handler` can be shared across to different
`thread-workers` resulting in multiple scheduler. So made sure that
only have **one scheduler**

 ## fixes done
* Updated readme
* replacing `if-let` / `when-let` with `some->`
* made sure of `scheduler` to be initialized only once and not any more